Add support for SD3.5
